#4 [feat] Product(엔티티) 수정
