--- god-care-api/src/main/java/com/godcare/api/service/Product.java
@@ -1,27 +1,0 @@
-package com.godcare.api.service;
-
-import com.godcare.common.dto.ResisterProductRequest;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@AllArgsConstructor
-@Getter
-public class Product {
-
-    private String mainImg;
-    private Long categoryId;
-    private Long id;
-
-    public void assignId(Long id) {
-        this.id = id;
-    }
-
-    public static Product toProduct(ResisterProductRequest request) {
-        return Product.builder()
-                .mainImg(request.getMainImg())
-                .categoryId(request.getCategoryId())
-                .build();
-    }
-}
